app/requirements.txt
fastapi
uvicorn[standard]
ultralytics
mlflow
python-multipart
python-dotenv
opencv-python
boto3